moved the home page to a separate component
